Move editor stuff to utils, add SuperTextarea to compat
