+++ js/src/forum/compat.js
@@ -73,7 +73,7 @@ import DiscussionListItem from './components/DiscussionListItem';
 import LoadingPost from './components/LoadingPost';
 import PostsUserPage from './components/PostsUserPage';
 import DiscussionPageResolver from './resolvers/DiscussionPageResolver';
-import BasicEditorDriver from './editor/BasicEditorDriver';
+import BasicEditorDriver from './utils/BasicEditorDriver';
 import routes from './routes';
 import ForumApplication from './ForumApplication';
 
@@ -86,6 +86,8 @@ export default Object.assign(compat, {
   'utils/alertEmailConfirmation': alertEmailConfirmation,
   'utils/UserControls': UserControls,
   'utils/Pane': Pane,
+  'utils/BasicEditorDriver': BasicEditorDriver,
+  'utils/SuperTextarea': BasicEditorDriver, // @deprecated beta 16, remove beta 17
   'states/ComposerState': ComposerState,
   'states/DiscussionListState': DiscussionListState,
   'states/GlobalSearchState': GlobalSearchState,
@@ -151,7 +153,6 @@ export default Object.assign(compat, {
   'components/LoadingPost': LoadingPost,
   'components/PostsUserPage': PostsUserPage,
   'resolvers/DiscussionPageResolver': DiscussionPageResolver,
-  'editors/BasicEditorDriver': BasicEditorDriver,
   routes: routes,
   ForumApplication: ForumApplication,
 });
