+++ js/src/forum/utils/EditorDriverInterface.ts
